requirements.txt
asgiref==3.8.1
Django==5.1.1
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.1
PyJWT==2.9.0
requests
sqlparse==0.5.1
django-cors-headers
razorpay
psycopg2-binary==2.9.9
python-dotenv
